Zeiter Teil Aufgabe 7 gemacht

+++ 0007_ARRAYS/ALGO_LIB_1/Exercise/lib.h
@@ -10,6 +10,6 @@ bool anyOf(int32_t *array, size_t length, int32_t value);
 
 bool noneOf(int32_t *array, size_t length, int32_t value);
 
-size_t count(int32_t *array, size_t length, int32_t value);
+long long unsigned int count(int32_t *array, size_t length, int32_t value);
 
 #endif
